Fixes comparison of DateTime fields during import
The `$newValue` for DateTime instances is normalized to a string, however, this is not done for `$existingValue`.  Therefore the comparison will always conclude that the value changed because `new \DateTime('2023-01-23') != '2023-01-23'`.

This bug causes all imports, that import any of the attributes, to always conclude that the data changed. This results in more database work and also a new revision for every entry on every import.