Rename mapping from parent to parentId
The attribute `parentId` is used in Craft to signify the value of the
parent element, if any.  Therefore, the mapping handle should also be
`parentId`.

This patch changes the mapping handle from `parent` into `parentId`.
This allows the comparison code to properly verify that an existing
parent already exists.  Without this patch the current comparison will
use the `parent` property that does not exist in a Craft entry.  This
should have been `parentId` which does exist.