Remove existing matrix blocks
If prepped data is empty then this returns an empty array. The effect being that any existing blocks will be removed.

This is, I believe, the expected result if one wants to removed already created blocks. Otherwise there does not seem to be a way to completely clear blocks from the Matrix.